feat: move util/verification to root shared folder
