+++ src/app/modules/verification/__tests__/verification.controller.spec.ts
@@ -12,11 +12,11 @@ import {
 } from 'src/app/services/sms/sms.errors'
 import { HashingError } from 'src/app/utils/hash'
 import * as OtpUtils from 'src/app/utils/otp'
-import { WAIT_FOR_OTP_SECONDS } from 'src/shared/util/verification'
 import { IFormSchema, IVerificationSchema } from 'src/types'
 
 import dbHandler from 'tests/unit/backend/helpers/jest-db'
 
+import { WAIT_FOR_OTP_SECONDS } from '../../../../../shared/util/verification'
 import expressHandler from '../../../../../tests/unit/backend/helpers/jest-express'
 import { DatabaseError, MalformedParametersError } from '../../core/core.errors'
 import { FormNotFoundError } from '../../form/form.errors'
