+++ src/app/modules/verification/verification.controller.ts
@@ -1,7 +1,7 @@
 import { celebrate, Joi, Segments } from 'celebrate'
 import { StatusCodes } from 'http-status-codes'
 
-import { SALT_ROUNDS } from '../../../shared/util/verification'
+import { SALT_ROUNDS } from '../../../../shared/util/verification'
 import { ErrorDto } from '../../../types/api'
 import { createLoggerWithLabel } from '../../config/logger'
 import { generateOtpWithHash } from '../../utils/otp'
