+++ src/app/services/mail/mail.service.ts
@@ -5,7 +5,7 @@ import Mail from 'nodemailer/lib/mailer'
 import promiseRetry from 'promise-retry'
 import validator from 'validator'
 
-import { HASH_EXPIRE_AFTER_SECONDS } from '../../../shared/util/verification'
+import { HASH_EXPIRE_AFTER_SECONDS } from '../../../../shared/util/verification'
 import {
   BounceType,
   EmailAdminDataField,
